test(frontend): tidy orchestration test splits
