+++ packages/frontend/src/state/operations/agent-orchestrator/events/session-errors-terminal.test.ts
@@ -1,62 +1,13 @@
 import { describe, expect, mock, test } from "bun:test";
-import { OPENCODE_RUNTIME_DESCRIPTOR } from "@openducktor/contracts";
-import { withMockedToast } from "@/test-utils/mock-toast";
 import {
-  lastSessionMessageForTest,
-  sessionMessageAt,
-  sessionMessagesToArray,
-} from "@/test-utils/session-message-test-helpers";
-import type { AgentSessionState } from "@/types/agent-orchestrator";
-import { createSessionEventBatcher } from "./session-event-batching";
-import type { SessionEvent, SessionPartEventContext } from "./session-event-types";
-import { attachAgentSessionListener, type SessionEventAdapter } from "./session-events";
-import { handleAssistantPart } from "./session-parts";
-
-const buildSession = (overrides: Partial<AgentSessionState> = {}): AgentSessionState => ({
-  runtimeKind: "opencode",
-  externalSessionId: "external-1",
-  taskId: "task-1",
-  repoPath: overrides.repoPath ?? "/tmp/repo",
-  role: "spec",
-  status: "running",
-  startedAt: "2026-02-22T08:00:00.000Z",
-  runtimeId: null,
-  workingDirectory: "/tmp/repo",
-  messages: [],
-  draftAssistantText: "",
-  draftAssistantMessageId: null,
-  draftReasoningText: "",
-  draftReasoningMessageId: null,
-  contextUsage: null,
-  pendingApprovals: [],
-  pendingQuestions: [],
-  todos: [],
-  modelCatalog: null,
-  selectedModel: null,
-  isLoadingModelCatalog: false,
-  ...overrides,
-});
-
-const getSession = (
-  sessionsRef: { current: Record<string, AgentSessionState> },
-  externalSessionId = "session-1",
-): AgentSessionState => {
-  const session = sessionsRef.current[externalSessionId];
-  if (!session) {
-    throw new Error(`Expected session ${externalSessionId}`);
-  }
-  return session;
-};
-
-const getSessionMessages = (
-  sessionsRef: { current: Record<string, AgentSessionState> },
-  externalSessionId = "session-1",
-) => sessionMessagesToArray(getSession(sessionsRef, externalSessionId));
-
-const getLastSessionMessage = (
-  sessionsRef: { current: Record<string, AgentSessionState> },
-  externalSessionId = "session-1",
-) => lastSessionMessageForTest(getSession(sessionsRef, externalSessionId));
+  type AgentSessionState,
+  attachAgentSessionListener,
+  buildSession,
+  getLastSessionMessage,
+  getSessionMessages,
+  OPENCODE_RUNTIME_DESCRIPTOR,
+  type SessionEventAdapter,
+} from "./session-events-test-harness";
 
 describe("agent-orchestrator session errors and terminal state", () => {
   test("keeps permission pending when auto-reject reply fails", async () => {
